integer because i cant be bothered to find how to gen a random float

+++ Horion/Module/Modules/Criticals.h
@@ -4,7 +4,7 @@ class Criticals :
 	public IModule
 {
 private:
-	float critChance = 100;
+	int critChance = 100;
 	
 public:
 	Criticals();
